# src/main/resources/swagger/gestaoobrasapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gest\u00e3o Obra API",
    "description": "API para gest\u00e3o de obras e controle de usu\u00e1rios.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/servico/status": {
      "post": {
        "tags": [
          "StatusServico"
        ],
        "operationId": "criarStatusServico",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusServicoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusServicoResponse"
                }
              }
            },
            "description": "Criado"
          }
        }
      },
      "get": {
        "tags": [
          "StatusServico"
        ],
        "operationId": "listarStatusServico",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusServicoResponse"
                  }
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/servico/status/{id}": {
      "get": {
        "tags": [
          "StatusServico"
        ],
        "operationId": "buscarStatusServicoPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusServicoResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "StatusServico"
        ],
        "operationId": "atualizarStatusServico",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusServicoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusServicoResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "StatusServico"
        ],
        "operationId": "excluirStatusServico",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Exclu\u00eddo"
          }
        }
      }
    },
    "/servico": {
      "post": {
        "tags": [
          "Servicos"
        ],
        "operationId": "criarServico",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicoResponse"
                }
              }
            },
            "description": "Criado"
          }
        }
      },
      "get": {
        "tags": [
          "Servicos"
        ],
        "operationId": "listarServicos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicoResponse"
                  }
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/servico/{id}": {
      "get": {
        "tags": [
          "Servicos"
        ],
        "operationId": "buscarServicoPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicoResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Servicos"
        ],
        "operationId": "atualizarServico",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicoResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Servicos"
        ],
        "operationId": "excluirServico",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Exclu\u00eddo"
          }
        }
      }
    },
    "/servico/material": {
      "post": {
        "tags": [
          "ServicoMaterial"
        ],
        "operationId": "criarServicoMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoMaterialRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicoMaterialResponse"
                }
              }
            },
            "description": "Criado"
          }
        }
      },
      "get": {
        "tags": [
          "ServicoMaterial"
        ],
        "operationId": "listarServicoMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicoMaterialResponse"
                  }
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/servico/material/{id}": {
      "get": {
        "tags": [
          "ServicoMaterial"
        ],
        "operationId": "buscarServicoMaterialPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicoMaterialResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "ServicoMaterial"
        ],
        "operationId": "atualizarServicoMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoMaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicoMaterialResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ServicoMaterial"
        ],
        "operationId": "excluirServicoMaterial",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Exclu\u00eddo"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "StatusServicoRequest": {
        "type": "object",
        "required": [
          "codigo",
          "nome"
        ],
        "properties": {
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          }
        }
      },
      "StatusServicoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ServicoRequest": {
        "type": "object",
        "required": [
          "idObra",
          "idStatusServico",
          "nome"
        ],
        "properties": {
          "idObra": {
            "type": "integer",
            "format": "int64"
          },
          "idStatusServico": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "observacao": {
            "type": "string"
          }
        }
      },
      "ServicoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idObra": {
            "type": "integer",
            "format": "int64"
          },
          "nomeObra": {
            "type": "string"
          },
          "idStatusServico": {
            "type": "integer",
            "format": "int64"
          },
          "nomeStatusServico": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "observacao": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ServicoMaterialRequest": {
        "type": "object",
        "required": [
          "idServico",
          "idObra",
          "idMaterial",
          "quantidade"
        ],
        "properties": {
          "idServico": {
            "type": "integer",
            "format": "int64"
          },
          "idObra": {
            "type": "integer",
            "format": "int64"
          },
          "idMaterial": {
            "type": "integer",
            "format": "int64"
          },
          "quantidade": {
            "type": "number"
          },
          "observacao": {
            "type": "string"
          }
        }
      },
      "ServicoMaterialResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idServico": {
            "type": "integer",
            "format": "int64"
          },
          "nomeServico": {
            "type": "string"
          },
          "idObra": {
            "type": "integer",
            "format": "int64"
          },
          "nomeObra": {
            "type": "string"
          },
          "idMaterial": {
            "type": "integer",
            "format": "int64"
          },
          "nomeMaterial": {
            "type": "string"
          },
          "quantidade": {
            "type": "number"
          },
          "observacao": {
            "type": "string"
          }
        }
      }
    }
  }
}